+++ lib/mongoid/validations.rb
@@ -1,4 +1,5 @@
 # encoding: utf-8
+require "mongoid/validations/macros"
 require "mongoid/validations/localizable"
 require "mongoid/validations/associated"
 require "mongoid/validations/format"
@@ -14,6 +15,11 @@ module Mongoid
   module Validations
     extend ActiveSupport::Concern
     include ActiveModel::Validations
+
+    included do
+      extend Macros
+      include Macros
+    end
 
     # Begin the associated validation.
     #
@@ -104,90 +110,19 @@ module Mongoid
 
     module ClassMethods
 
-      # Validates whether or not an association is valid or not. Will correctly
-      # handle has one and has many associations.
+      # Adds an associated validator for the relation if the validate option
+      # was not provided or set to true.
       #
-      # @example
+      # @example Set up validation.
+      #   Person.validates_relation(metadata)
       #
-      #   class Person
-      #     include Mongoid::Document
-      #     embeds_one :name
-      #     embeds_many :addresses
+      # @param [ Metadata ] metadata The relation metadata.
       #
-      #     validates_associated :name, :addresses
-      #   end
-      #
-      # @param [ Array ] *args The arguments to pass to the validator.
-      def validates_associated(*args)
-        validates_with(AssociatedValidator, _merge_attributes(args))
-      end
-
-      # Validates whether or not a field is unique against the documents in the
-      # database.
-      #
-      # @example
-      #
-      #   class Person
-      #     include Mongoid::Document
-      #     field :title
-      #
-      #     validates_uniqueness_of :title
-      #   end
-      #
-      # @param [ Array ] *args The arguments to pass to the validator.
-      def validates_uniqueness_of(*args)
-        validates_with(UniquenessValidator, _merge_attributes(args))
-      end
-
-      # Validates the format of a field.
-      #
-      # @example
-      #   class Person
-      #     include Mongoid::Document
-      #     field :title
-      #
-      #     validates_format_of :title, with: /^[a-z0-9 \-_]*$/i
-      #   end
-      #
-      # @param [ Array ] args The names of the fields to validate.
-      #
-      # @since 2.4.0
-      def validates_format_of(*args)
-        validates_with(Mongoid::Validations::FormatValidator, _merge_attributes(args))
-      end
-
-      # Validates the length of a field.
-      #
-      # @example
-      #   class Person
-      #     include Mongoid::Document
-      #     field :title
-      #
-      #     validates_length_of :title, minimum: 100
-      #   end
-      #
-      # @param [ Array ] args The names of the fields to validate.
-      #
-      # @since 2.4.0
-      def validates_length_of(*args)
-        validates_with(Mongoid::Validations::LengthValidator, _merge_attributes(args))
-      end
-
-      # Validates whether or not a field is present - meaning nil or empty.
-      #
-      # @example
-      #   class Person
-      #     include Mongoid::Document
-      #     field :title
-      #
-      #     validates_presence_of :title
-      #   end
-      #
-      # @param [ Array ] args The names of the fields to validate.
-      #
-      # @since 2.4.0
-      def validates_presence_of(*args)
-        validates_with(PresenceValidator, _merge_attributes(args))
+      # @since 2.0.0.rc.1
+      def validates_relation(metadata)
+        if metadata.validate?
+          validates_associated(metadata.name)
+        end
       end
 
       # Add validation with the supplied validators forthe provided fields
@@ -213,23 +148,6 @@ module Mongoid
           end
         end
         super
-      end
-
-      private
-
-      # Adds an associated validator for the relation if the validate option
-      # was not provided or set to true.
-      #
-      # @example Set up validation.
-      #   Person.validates_relation(metadata)
-      #
-      # @param [ Metadata ] metadata The relation metadata.
-      #
-      # @since 2.0.0.rc.1
-      def validates_relation(metadata)
-        if metadata.validate?
-          validates_associated(metadata.name)
-        end
       end
 
       # Are we currently performing a validation that has a query?